Adding Authorization Module

+++ Restaurant.Application/Dishes/Commands/DeleteDish/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurant.Domain.Constants;
 using Restaurant.Domain.Entities;
 using Restaurant.Domain.Exceptions;
+using Restaurant.Domain.Interfaces;
 using Restaurant.Domain.Repsitories;
 using System;
 using System.Collections.Generic;
@@ -13,7 +15,7 @@ namespace Restaurant.Application.Dishes.Commands.DeleteDish
 {
     public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger,
     IResutaurantRepository restaurantsRepository,
-    IDishRepository dishesRepository)
+    IDishRepository dishesRepository, IRestaurantAuthorizationService restaurantAuthorizationService)
         : IRequestHandler<DeleteDishesForRestaurantCommand>
     {
         public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
@@ -22,6 +24,9 @@ namespace Restaurant.Application.Dishes.Commands.DeleteDish
             var restauarnt = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
             if(restauarnt == null)
                 throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+            
+            if (!restaurantAuthorizationService.Authorize(restauarnt, ResourceOperation.Update))
+                throw new ForbidException();
 
             await dishesRepository.Delete(restauarnt.Dishes);
         }
